+++ WelcomeHome/WelcomeHome.DAL/Migrations/20240204180047_EventModelChanged.cs
@@ -10,44 +10,44 @@ namespace WelcomeHome.DAL.Migrations
         /// <inheritdoc />
         protected override void Up(MigrationBuilder migrationBuilder)
         {
-            migrationBuilder.AlterColumn<int>(
+            migrationBuilder.AlterColumn<long>(
                 name: "VolunteerId",
                 table: "Events",
-                type: "int",
+                type: "bigint",
                 nullable: true,
-                oldClrType: typeof(int),
-                oldType: "int");
+                oldClrType: typeof(long),
+                oldType: "bigint");
 
-            migrationBuilder.AlterColumn<int>(
+            migrationBuilder.AlterColumn<long>(
                 name: "EstablishmentId",
                 table: "Events",
-                type: "int",
+                type: "bigint",
                 nullable: true,
-                oldClrType: typeof(int),
-                oldType: "int");
+                oldClrType: typeof(long),
+                oldType: "bigint");
         }
 
         /// <inheritdoc />
         protected override void Down(MigrationBuilder migrationBuilder)
         {
-            migrationBuilder.AlterColumn<int>(
+            migrationBuilder.AlterColumn<long>(
                 name: "VolunteerId",
                 table: "Events",
-                type: "int",
+                type: "bigint",
                 nullable: false,
                 defaultValue: 0,
-                oldClrType: typeof(int),
-                oldType: "int",
+                oldClrType: typeof(long),
+                oldType: "bigint",
                 oldNullable: true);
 
-            migrationBuilder.AlterColumn<int>(
+            migrationBuilder.AlterColumn<long>(
                 name: "EstablishmentId",
                 table: "Events",
-                type: "int",
+                type: "bigint",
                 nullable: false,
                 defaultValue: 0,
-                oldClrType: typeof(int),
-                oldType: "int",
+                oldClrType: typeof(long),
+                oldType: "bigint",
                 oldNullable: true);
         }
     }
