+++ WelcomeHome/WelcomeHome.DAL/Models/Establishment.cs
@@ -7,11 +7,11 @@ namespace WelcomeHome.DAL.Models
     public class Establishment
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; } = null!;
         public string Address { get; set; } = null!;
-        public int EstablishmentTypeId { get; set; }
-        public int CityId { get; set; }
+        public long EstablishmentTypeId { get; set; }
+        public long CityId { get; set; }
         public string? PageURL { get; set; }
         public string? PhoneNumber { get; set; }
         public string? OtherContacts { get; set; }
