+++ WelcomeHome/WelcomeHome.DAL/Models/RefreshToken.cs
@@ -2,9 +2,9 @@
 
 public class RefreshToken
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Token { get; set; }
     public DateTime Expires { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; }
 }
