+++ WelcomeHome/WelcomeHome.DAL/Models/Volunteer.cs
@@ -5,10 +5,10 @@ namespace WelcomeHome.DAL.Models
     public class Volunteer
     {
         [Key]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
         public string SocialUrl { get; set; }
-        public int? EstablishmentId { get; set; }
+        public long? EstablishmentId { get; set; }
         public Establishment? Establishment { get; set; }
         public ICollection<Event>? Events { get; set; }
     }
