+++ WelcomeHome/WelcomeHome.DAL/Repositories/CityRepository.cs
@@ -20,7 +20,7 @@ public sealed class CityRepository : ICityRepository
                        .Select(c => c);
     }
 
-    public async Task<City?> GetByIdAsync(int id)
+    public async Task<City?> GetByIdAsync(long id)
     {
         return await _context.Cities
                              .Include(c => c.Country)
@@ -45,7 +45,7 @@ public sealed class CityRepository : ICityRepository
         await _context.SaveChangesAsync().ConfigureAwait(false);
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task DeleteAsync(long id)
     {
         var foundCity = await _context.Cities.SingleAsync(c => c.Id == id).ConfigureAwait(false);
