+++ WelcomeHome/WelcomeHome.DAL/Repositories/CountryRepository.cs
@@ -20,7 +20,7 @@ public sealed class CountryRepository : ICountryRepository
                        .Select(c => c);
     }
 
-    public async Task<Country?> GetByIdAsync(int id)
+    public async Task<Country?> GetByIdAsync(long id)
     {
         return await _context.Countries
                              .Include(c => c.Cities)
@@ -41,7 +41,7 @@ public sealed class CountryRepository : ICountryRepository
         await _context.SaveChangesAsync().ConfigureAwait(false);
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task DeleteAsync(long id)
     {
         var foundCountry = await _context.Countries.SingleAsync(c => c.Id == id).ConfigureAwait(false);
