+++ WelcomeHome/WelcomeHome.DAL/Repositories/CourseRepository.cs
@@ -17,7 +17,7 @@ namespace WelcomeHome.DAL.Repositories
             return _context.Courses.Select(e => e);
         }
 
-        public async Task<Course?> GetByIdAsync(int id)
+        public async Task<Course?> GetByIdAsync(long id)
         {
             return await _context.Courses.FirstOrDefaultAsync(e => e.Id == id).ConfigureAwait(false);
         }
@@ -29,7 +29,7 @@ namespace WelcomeHome.DAL.Repositories
             await _context.SaveChangesAsync().ConfigureAwait(false);
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task DeleteAsync(long id)
         {
             var existingCourse = await _context.Courses.SingleAsync(c => c.Id == id).ConfigureAwait(false);
             _context.Courses.Remove(existingCourse);
